Change DirByTime to try to quite earlier, from cleanup task

+++ storemate-store/src/main/java/com/fasterxml/storemate/store/file/DirByTime.java
@@ -1,6 +1,7 @@
 package com.fasterxml.storemate.store.file;
 
 import java.io.*;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.joda.time.DateTime;
 
@@ -26,10 +27,10 @@ public class DirByTime
      * Method called to delete this directory along with all of
      * its contents.
      */
-    public int nuke(FileCleanupStats stats) {
-        return _nukeDir(_dir, stats);
+    public int nuke(FileCleanupStats stats, AtomicBoolean shutdown) {
+        return _nukeDir(_dir, stats, shutdown);
     }
-
+    
     /**
      * Method called to find and delete all empty directories
      * within this directory; as well as directory itself
@@ -37,18 +38,21 @@ public class DirByTime
      * 
      * @return True if this directory was empty and removed.
      */
-    public boolean removeEmpty(FileCleanupStats stats) {
-        return _removeEmpty(_dir, stats);
+    public boolean removeEmpty(FileCleanupStats stats, AtomicBoolean shutdown) {
+        return _removeEmpty(_dir, stats, shutdown);
     }
     
-    protected int _nukeDir(File dir, FileCleanupStats stats)
+    protected int _nukeDir(File dir, FileCleanupStats stats, AtomicBoolean shutdown)
     {
         int failed = 0;
         int childCount = 0;
         for (File f : dir.listFiles()) {
+            if (shutdown != null && shutdown.get()) {
+                return failed;
+            }
             ++childCount;
             if (f.isDirectory()) {
-                failed += _nukeDir(f, stats);
+                failed += _nukeDir(f, stats, shutdown);
             } else {
                 if (f.delete()) {
                     stats.addDeletedFile();
@@ -71,12 +75,15 @@ public class DirByTime
         return failed;
     }
 
-    protected boolean _removeEmpty(File dir, FileCleanupStats stats)
+    protected boolean _removeEmpty(File dir, FileCleanupStats stats, AtomicBoolean shutdown)
     {
         int remaining = 0;
         for (File f : dir.listFiles()) {
+            if (shutdown != null && shutdown.get()) {
+                break;
+            }
             if (f.isDirectory()) {
-                if (_removeEmpty(f, stats)) {
+                if (_removeEmpty(f, stats, shutdown)) {
                     stats.addDeletedEmptyDir();
                     continue;
                 }
